[refactor] Move IRBank to a separate file
